optimization of parallel run
