+++ src/test/java/com/softserve/edu/greencity/ui/tests/runner/GreenCityTestRunner.java
@@ -80,7 +80,7 @@ public abstract class GreenCityTestRunner {
                     new URL("http://localhost:4444/wd/hub"), options);
         }
         driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
-        driver.manage().timeouts().pageLoadTimeout(65, TimeUnit.SECONDS);
+        driver.manage().timeouts().pageLoadTimeout(80, TimeUnit.SECONDS);
         driver.manage().window().maximize();
         /*<============================Local============================>*/
     }
