+++ testng.xml
@@ -1,55 +1,50 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Suite" parallel="classes" thread-count="5">
-<!-- ===============================   UI  ============================= -->
-    <test name="UI_smoke">
+
+    <test name="Quick_tests" thread-count="6">
+    <!-- API and other tests that run quickly (1 minute and less) -->
         <classes>
             <class name="com.softserve.edu.greencity.ui.tests.SmokeTest"/>
 
+            <class name="com.softserve.edu.greencity.api.tests.econews.positive.CreateNewsWithMinimalValuesTest"/>
+            <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithArrayResponseTest"/>
+            <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithDetailedResponseTest"/>
+            <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithShortResponseTest"/>
+
+            <class name="com.softserve.edu.greencity.api.tests.signin.SignInApiTest"/>
         </classes>
     </test>
 
-    <test name="UI_registrationAndLogin">
+    <test name="Registration_and_login">
         <classes>
             <class name="com.softserve.edu.greencity.ui.tests.signup.RegisterPageTests"/>
             <class name="com.softserve.edu.greencity.ui.tests.signup.RegistrationTests"/>
             <class name="com.softserve.edu.greencity.ui.tests.signin.ForgotPasswordTests"/>
             <class name="com.softserve.edu.greencity.ui.tests.signin.LoginTest"/>
-<!--
-            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsListViewTests">
-                <methods>
-                    <include name="isItemsDisplayedChronological" />
-                </methods>
-            </class>
--->
+
+            <class name="com.softserve.edu.greencity.ui.tests.viewsinglenews.EcoNewsSingleViewTest"/>
+            <!-- EcoNewsSingleViewTest needs no another news to be created, but creates by itself.
+            So it could be merged only with this suite -->
         </classes>
     </test>
 
-    <test name="UI_createNews">
+    <test name="Google_login" thread-count="1">
+    <!-- Should be run separately, since it interacts with other windows (Google login window) -->
+        <classes>
+            <class name="com.softserve.edu.greencity.ui.tests.signin.GoogleSignInTest"/>
+        </classes>
+    </test>
+
+    <test name="Create_news_independent" thread-count="4">
+    <!-- Tests that create news, but will not fail if other news is created at the same time
+    (these tests should use switchToSingleNewsPageByParameters, but not ByNumber) -->
         <classes>
             <class name="com.softserve.edu.greencity.ui.tests.createnews.CreateNewsPositiveTest"/>
             <class name="com.softserve.edu.greencity.ui.tests.createnews.CreateNewsNegativeTest"/>
             <class name="com.softserve.edu.greencity.ui.tests.createnews.CreateNewsPreviewTest"/>
             <class name="com.softserve.edu.greencity.ui.tests.createnews.CreateNewsLegacyTest"/>
-        </classes>
-    </test>
 
-    <test name="UI_viewAllNews">
-        <classes>
-            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsGridViewTest"/>
-            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsListViewTests"/>
-<!--            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsPageTest"/>-->
-        </classes>
-    </test>
-
-    <test name="UI_viewSingleNews">
-        <classes>
-            <class name="com.softserve.edu.greencity.ui.tests.viewsinglenews.EcoNewsSingleViewTest"/>
-        </classes>
-    </test>
-
-    <test name="UI_comments">
-        <classes>
             <class name="com.softserve.edu.greencity.ui.tests.comments.CheckElementOfCommentTest"/>
             <class name="com.softserve.edu.greencity.ui.tests.comments.CommentCreation"/>
             <class name="com.softserve.edu.greencity.ui.tests.comments.Commentstest"/>
@@ -57,20 +52,22 @@
         </classes>
     </test>
 
-<!-- ===============================   API  ============================= -->
-    <test name="API_registrationAndLogin">
+    <test name="View_all_news">
+<!--        Should be run separately from the tests that create or delete news. Don't create news by themselves-->
         <classes>
-            <class name="com.softserve.edu.greencity.api.tests.signin.SignInApiTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsGridViewTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsListViewTests"/>
+<!--            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsPageTest"/>-->
         </classes>
     </test>
 
-    <test name="API_createNews">
-        <classes>
-            <class name="com.softserve.edu.greencity.api.tests.econews.positive.CreateNewsWithMinimalValuesTest"/>
-            <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithArrayResponseTest"/>
-            <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithDetailedResponseTest"/>
-            <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithShortResponseTest"/>
-        </classes>
-    </test>
+
+    <!--
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsListViewTests">
+                <methods>
+                    <include name="isItemsDisplayedChronological" />
+                </methods>
+            </class>
+-->
 
 </suite>
